+++ frontend/test/__runner__/run_cypress_tests.js
@@ -2,7 +2,7 @@ const { printBold } = require("./cypress-runner-utils");
 const runCypress = require("./cypress-runner-run-tests");
 const getVersion = require("./cypress-runner-get-version");
 const generateSnapshots = require("./cypress-runner-generate-snapshots");
-const CypressBackend = require("./backend.js");
+const CypressBackend = require("./cypress-runner-backend");
 
 const e2eHost = process.env["E2E_HOST"];
